Fix release build target not buildable
This commit fixes that the release build type/target was not buildable
caused by missconfigured `CMakeLists.txt` as the variable `TYPE` was
empty in the `release` target.

No `TYPE` is an option that defaults to `release` making its built
possible.

Additionally added a new workflow that builds all tools, without running
them