Remove blank lines from git cmd strm in `test_githlpr`
This commit removes unused blank lines in the `test_githlpr` unit tests,
as these are not needed. Git does not terminate its commands with a
blank line.

These blank lines were a relic from older version where it was assumed
that Git also terminated its cmds with a blank line.
Removed it as not needed and does not represent actual Git behavior