Fix git cmd stream blank line triggers blank line output
This commit fixes an erronous merge that made blank-lines in the git cmd
stream to print blank lines in the reply.
This made test fail.

I should not merge unfinished changes onto branch as it was not easy to
debug as other tests were failing on purpose and the blank line
regression was not caught.
So next time only merge finished changes onto master